Adding Navigator.waitForPendingNavigations() to enable some level of chaining (#3)
Summary:

At Geekie, we have one particular use case in which one screen may "redirect" to another screen (via navigator.replace). Sometimes, this redirect happens as soon as the first screen mounts, and in this case there is a race condition: the call to replace() may happen before the lock is released, causing the redirect to be silently ignored.

This commit exposes a waitForPendingNavigations() API that allows navigator clients to await for the current transition to finish prior to issuing new navigation commands. This is insufficient to implement a true queueing mechanism, but then there seems to be no currently use cases that would require a true queueing system.

Test Plan:

- Used over Geekie to fix a bug in the HighlightOpener screen
- Ran `npx jest`
- Ran `npx eslint` over all *.js files